pyproject.toml
[project]
name = "ox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.1.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]
